Relat-3.0.5/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
requests==2.26.0
mysql-connector-python==8.0.26
beautifulsoup4==4.10.0
